use service worker from Create React App PWA template
