+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 
 import App from './app'
-import * as serviceWorker from './service-worker'
+import * as serviceWorkerRegistration from './serviceWorkerRegistration'
 
 import './styling/index.scss'
 
@@ -11,4 +11,4 @@ ReactDOM.render(<App />, document.getElementById('root'))
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister()
+serviceWorkerRegistration.unregister()
